streamlit/requirements.txt
streamlit
pandas
plotly
clickhouse-connect
psycopg2-binary
sentence-transformers
torch